+++ pyo3cls/src/lib.rs
@@ -5,8 +5,8 @@
 extern crate proc_macro;
 use proc_macro::TokenStream;
 use pyo3_derive_backend::{
-    build_py_class, build_py_function, build_py_methods, build_py_proto, get_doc,
-    process_functions_in_module, py_init, PyClassArgs, PyFunctionAttr,
+    build_derive_from_pyobject, build_py_class, build_py_function, build_py_methods,
+    build_py_proto, get_doc, process_functions_in_module, py_init, PyClassArgs, PyFunctionAttr,
 };
 use quote::quote;
 use syn::parse_macro_input;
@@ -87,6 +87,16 @@ pub fn pyfunction(attr: TokenStream, input: TokenStream) -> TokenStream {
 
     quote!(
         #ast
+        #expanded
+    )
+    .into()
+}
+
+#[proc_macro_derive(FromPyObject, attributes(pyo3, extract))]
+pub fn derive_from_py_object(item: TokenStream) -> TokenStream {
+    let ast = parse_macro_input!(item as syn::DeriveInput);
+    let expanded = build_derive_from_pyobject(&ast).unwrap_or_else(|e| e.to_compile_error());
+    quote!(
         #expanded
     )
     .into()